# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CaReMind API
  version: 1.0.0
  description: REST API for private vehicle, maintenance and operating-cost management.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.car-remind.gr/api
    description: Production
...
paths:
...
  /maintenances:
    get:
      operationId: listMaintenances
      tags: [Maintenance]
      summary: List maintenance records
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Maintenance list }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createMaintenance
      tags: [Maintenance]
      summary: Add maintenance for an owned vehicle
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/Maintenance' }
      responses:
        '201': { description: Maintenance created }
        '404': { description: Vehicle is not owned by this user }
  /maintenances/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      operationId: updateMaintenance
      tags: [Maintenance]
      summary: Update an owned maintenance record
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/Maintenance' }
      responses:
        '200': { description: Maintenance updated }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteMaintenance
      tags: [Maintenance]
      summary: Delete an owned maintenance record
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Maintenance deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /costs:
    get:
      operationId: listCosts
      tags: [Costs]
      summary: List cost entries
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Cost list }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createCost
      tags: [Costs]
      summary: Add a cost to an owned vehicle
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/Cost' }
      responses:
        '201': { description: Cost created }
        '404': { description: Vehicle is not owned by this user }
  /costs/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      operationId: updateCost
      tags: [Costs]
      summary: Update an owned cost
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/Cost' }
      responses:
        '200': { description: Cost updated }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteCost
      tags: [Costs]
      summary: Delete an owned cost
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Cost deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /cron/maintenance:
    get:
      operationId: runMaintenanceReminders
      tags: [Automation]
      summary: Send due maintenance reminders
      security: [{ cronSecret: [] }]
      parameters:
        - in: header
          name: X-Cron-Secret
          required: true
          schema: { type: string }
      responses:
        '200': { description: Reminder run completed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { description: Cron service is not configured }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cronSecret:
      type: apiKey
      in: header
      name: X-Cron-Secret
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
...
    Unauthorized:
      description: Missing, invalid or expired authentication
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found or not owned by this user
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  requestBodies:
...
    Maintenance:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Maintenance' }
    Cost:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Cost' }
  schemas:
    Error:
      type: object
      properties: { error: { type: string }, code: { type: string } }
...
    Maintenance:
      type: object
      required: [vehicleId, maintenanceType]
      properties:
        id: { type: integer, readOnly: true }
        vehicleId: { type: integer, minimum: 1 }
        maintenanceType: { type: string, maxLength: 100 }
        lastDate: { type: [string, 'null'], format: date }
        nextDate: { type: [string, 'null'], format: date }
        lastMileage: { type: [integer, 'null'], minimum: 0 }
        nextMileage: { type: [integer, 'null'], minimum: 0 }
        notificationDays: { type: integer, minimum: 0, maximum: 365, default: 7 }
        status: { type: string, enum: [active, pending, completed, overdue] }
        notes: { type: [string, 'null'] }
    Cost:
      type: object
      required: [vehicleId, category, amount, date]
      properties:
        id: { type: integer, readOnly: true }
        vehicleId: { type: integer, minimum: 1 }
        category: { type: string, maxLength: 100 }
        amount: { type: number, exclusiveMinimum: 0 }
        date: { type: string, format: date }
        description: { type: [string, 'null'] }
        receiptNumber: { type: [string, 'null'], maxLength: 100 }